apps/api/requirements.txt
fastapi
uvicorn
whoosh
python-multipart
pydantic
pandas
PySastrawi
